- Moved Flink turbine code to `iot` package to mirror the standalone code. - Used DNS name for Pravega controller endpoint. - Remove dependency on lombok. - Updated README
