+++ flink-examples/src/main/scala/com/emc/pravega/examples/flink/util/PravegaParameters.scala
@@ -3,7 +3,7 @@
   * Copyright (c) 2017 Dell Inc., or its subsidiaries.
   *
   */
-package com.emc.pravega.examples.flink
+package com.emc.pravega.examples.flink.util
 
 object PravegaParameters {
   val DEFAULT_CONTROLLER_URI = "tcp://controller-pravega.marathon.mesos:9091"
